+++ lib/src/youtube_iframe/youtube_iframe.dart
@@ -12,9 +12,11 @@ import '../tube_service.dart';
   styleUrls: const ['youtube_iframe.css'],
 )
 class YoutubeIFrame implements OnInit {
+  YoutubeIFrame(this.ytService);
+
   TubeService ytService;
 
-  YoutubeIFrame(this.ytService);
+  PlayerState currentState = PlayerState.notStarted;
 
   @Input()
   int width = 480;
@@ -23,39 +25,112 @@ class YoutubeIFrame implements OnInit {
   int height = 360;
 
   @Input()
-  String videoId;
+  void set videoId(String value) {
+    _videoId = value;
+    updatePlayer();
+  }
+
+  String _videoId;
 
   @Input()
   String playerId = 'ytframe';
 
   @Input()
-  bool showControls = true;
-  int get controls => showControls ? 1 : 0;
+  void set showControls(bool value) {
+    print('YoutubeIFrame.showControls... $value');
+    _showControls = value;
+    updatePlayer();
+  }
+
+  bool _showControls = true;
+  int get controls => _showControls ? 1 : 0;
 
   @Input()
-  bool showInfos = false;
-  int get infos => showInfos ? 1 : 0;
+  void set showInfos(bool value) {
+    _showInfos = value;
+    updatePlayer();
+  }
+
+  bool _showInfos = false;
+  int get infos => _showInfos ? 1 : 0;
+
+  bool _autoPlay = false;
 
   @Input()
-  bool autoPlay = false;
-  int get willPlayAuto => autoPlay ? 1 : 0;
+  void set autoPlay(bool value) {
+    _autoPlay = value;
+    updatePlayer();
+  }
+
+  int get willPlayAuto => _autoPlay ? 1 : 0;
 
   @Input()
-  bool disableKeyboard = false;
-  int get disablekb => autoPlay ? 1 : 0;
+  void set disableKeyboard(bool value) {
+    _disableKeyboard = value;
+    updatePlayer();
+  }
+
+  bool _disableKeyboard = false;
+  int get disablekb => _disableKeyboard ? 1 : 0;
 
   @Input()
-  bool embedOnMobile = false;
-  int get playsinline => embedOnMobile ? 1 : 0;
+  void set modestBranding(bool value) {
+    _modestBranding = value;
+    updatePlayer();
+  }
 
-  StreamController<Null> readyStreamer = new StreamController<Null>();
+  bool _modestBranding = false;
+  int get modestbranding => _modestBranding ? 1 : 0;
+
+  @Input()
+  void set mobilePlaysInline(bool value) {
+    _mobilePlaysInline = value;
+    updatePlayer();
+  }
+
+  bool _mobilePlaysInline = false;
+  int get playsinline => _mobilePlaysInline ? 1 : 0;
+
+  @Input()
+  void set allowFS(bool value) {
+    _allowFS = value;
+    updatePlayer();
+  }
+
+  bool _allowFS = false;
+  int get fs => _allowFS ? 1 : 0;
+
+  @Input()
+  void set start(int value) {
+    _start = value;
+    updatePlayer();
+  }
+
+  int _start = 0;
+
+  @Input()
+  void set end(int value) {
+    _end = value;
+    updatePlayer();
+  }
+
+  int _end = 0;
+
+  StreamController<Player> readyStreamer =
+      new StreamController<Player>.broadcast();
   @Output('onReady')
-  Stream<Null> get ready$ => readyStreamer.stream;
+  Stream<Player> get ready$ => readyStreamer.stream;
 
   StreamController<PlayerState> playerStateStreamer =
-      new StreamController<PlayerState>();
+      new StreamController<PlayerState>.broadcast();
   @Output('onStateChange')
   Stream<PlayerState> get playerState$ => playerStateStreamer.stream;
+
+  StreamController<YTPlayerError> errorStreamer =
+      new StreamController<YTPlayerError>.broadcast();
+
+  @Output('onError')
+  Stream<YTPlayerError> get error$ => errorStreamer.stream;
 
   Player player;
 
@@ -70,7 +145,7 @@ class YoutubeIFrame implements OnInit {
         new PlayerOptions(
             width: width.toString(),
             height: height.toString(),
-            videoId: videoId,
+            videoId: _videoId,
             playerVars: new PlayerVars(
               controls: controls,
               showinfo: infos,
@@ -78,16 +153,32 @@ class YoutubeIFrame implements OnInit {
               autoplay: willPlayAuto,
               enablejsapi: 1,
               playsinline: playsinline,
+              fs: fs,
+              start: _start,
+              end: _end,
             ),
-            events:
-                new Events(onReady: onReady, onStateChange: onStateChange)));
+            events: new Events(
+                onReady: onReady,
+                onStateChange: onStateChange,
+                onError: (EventArgs ev) =>
+                    errorStreamer.add(getError(ev.data)))));
+    playerState$.listen((s) => currentState = s);
   }
 
   void onReady(EventArgs eventArgs) {
-    readyStreamer.add(null);
+    readyStreamer.add(player);
   }
 
   void onStateChange(EventArgs eventArgs) {
     playerStateStreamer.add(stateFromInt(eventArgs.data));
+  }
+
+  void updatePlayer() {
+    if (player != null) {
+      if( currentState == PlayerState.playing)
+        player.stopVideo();
+      player.destroy();
+      createPlayer();
+    }
   }
 }